feat(postges): support postgres repo --> db
- add postgres migration for Todos
- add databases/postgres.go for implementing CRUD

+++ internal/modules/todo/service.go
@@ -1,6 +1,9 @@
 package todo
 
-import "fmt"
+import (
+	"fmt"
+	"golang-api-starter/internal/helper"
+)
 
 type Service struct {
 	repo *Repository
@@ -15,6 +18,16 @@ func (s *Service) Get(queries map[string]interface{}) []*Todo {
 	return s.repo.Get(queries)
 }
 
+func (s *Service) GetById(queries map[string]interface{}) ([]*Todo, error) {
+	fmt.Printf("todo service\n")
+
+	records := s.repo.Get(queries)
+	if len(records) == 0 {
+		return nil, fmt.Errorf("%s with id: %s not found", tableName, queries["id"])
+	}
+	return records, nil
+}
+
 func (s *Service) Create(todos []*Todo) []*Todo {
 	fmt.Printf("todo service create\n")
 	return s.repo.Create(todos)
@@ -26,5 +39,13 @@ func (s *Service) Update(todos []*Todo) []*Todo {
 }
 
 func (s *Service) Delete(ids *[]int64) ([]*Todo, error) {
+	idsString, _ := helper.ConvertNumberSliceToString(*ids)
+	records := s.repo.Get(map[string]interface{}{
+		"id": idsString,
+	})
+	if len(records) == 0 {
+		return nil, fmt.Errorf("failed to delete, %s with id: %+v not found", tableName, ids)
+	}
+
 	return s.repo.Delete(ids)
 }
